Higher order component that adds style classes to the outer block wrapper on init and change in the editor
