+++ wp-content/themes/core/assets/js/src/admin/editor/index.js
@@ -1,24 +1,5 @@
-import { getBlockTypes, unregisterBlockType } from '@wordpress/blocks';
-import { BLOCK_BLACKLIST } from '../config/wp-settings';
-
-/**
- * @function removeBlackListedBlocks
- * @description Takes an array supplied on our config object and unregisters those blocks from Gutenberg after first
- * checking that they are registered in the current admin context
- */
-
-const removeBlackListedBlocks = () => {
-	const registeredBlockTypes = getBlockTypes().map( block => block.name );
-	const blocksToUnregister = BLOCK_BLACKLIST.filter( blockName => registeredBlockTypes.includes( blockName ) );
-
-	if ( ! blocksToUnregister.length ) {
-		return;
-	}
-
-	blocksToUnregister.forEach( type => unregisterBlockType( type ) );
-
-	console.info( 'SquareOne BE: Unregistered these blocks from Gutenberg: ', blocksToUnregister );
-};
+import hooks from './hooks';
+import types from './types';
 
 /**
  * @function init
@@ -26,7 +7,8 @@ const removeBlackListedBlocks = () => {
  */
 
 const init = () => {
-	removeBlackListedBlocks();
+	hooks();
+	types();
 };
 
 export default init;
